+++ commands/fun.py
@@ -2,46 +2,46 @@
 import discord
 from discord.ext import commands
 from services import gifs
-from services.homura import mplayer 
-from services.homura.permission import has_permission
-from services.homura.permission import has_prefix_permission
+from services.homura import player 
+from services.homura.MediaLoader import ydl_clear_cache
+from services.homura.permission import has_permission, has_prefix_permission
 
 class Fun(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
-        mplayer.ydl_clear_cache()
+        ydl_clear_cache()
 
     # Audio/Music command field
 
     @commands.command(help='Play audio on a voice channel.')
     @has_prefix_permission()
     async def playm(self, ctx, *, search):
-        await mplayer.music_player(ctx, search)
+        await player.audio_player(ctx, search)
 
     @commands.command(help='Remove specifict audio from queue.')
     @has_prefix_permission()
     async def removem(self, ctx, *, query):
-        await mplayer.remove_queue(ctx, query)
+        await player.remove_queue(ctx, query)
 
     @commands.command(help='Clear the queue, keep playing the current audio.')
     @has_prefix_permission()
     async def clearm(self, ctx):
-        await mplayer.clear_whole_queue(ctx)
+        await player.clear_whole_queue(ctx)
 
     @commands.command(help='Skip one or more queue(s).')
     @has_prefix_permission()
     async def skipm(self, ctx, num='1', skip_current=None):
-        await mplayer.clear_queue(ctx, num, skip_current)
+        await player.clear_queue(ctx, num, skip_current)
 
     @commands.command(help='Show the details of current queue.')
     @has_prefix_permission()
     async def queue(self, ctx):
-        await mplayer.show_queue(self.bot, ctx)
+        await player.show_queue(self.bot, ctx)
 
     @commands.command(help='Stop the audio and leave the voice channel.')
     @has_prefix_permission()
     async def stopm(self, ctx):
-        await mplayer.music_stop(ctx)
+        await player.audio_stop(ctx)
 
     # GIFs command field