+++ config.py
@@ -4,11 +4,12 @@ import shutil
 
 @dataclass(frozen=True)
 class Config:
-    VERSION = "0.6.3"
+    VERSION = "0.6.4"
     BOT_NAME = "Homura"
     CREATOR = "Kanade"
     PREFIX = "$"
     FFMPEG_PATH = os.getenv("FFMPEG_PATH") or shutil.which("ffmpeg") or os.path.join(os.path.dirname(os.path.abspath(__file__)), "tools/ffmpeg/bin", "ffmpeg")
+    DENO_PATH = "tools/deno/deno"
     MASTER = 123456789012345678
     SAWERIA = "https://saweria.co/kanaede"