+++ changelog.md
@@ -1,5 +1,12 @@
-# Changelog 0.6.3
+# Changelog 0.6.4
 
-## Bugs fixed
+- yt-dlp updated to 2026.08.19
+    - fixed 403 error forbidden
+- refactor mplayer
+- rename database `guildMusic` to `guildAudio`
+- deno path as config
 
-- fixed audio won't play using youtube url
+## Bugs Fixed
+
+- fixed residue from past git conflict
+- fixed news field showed up even when `news.md` is empty